extract CustomShapeBlock
